Rename ArticlePlugin to ArticlePackage.
While this is implemented like a plugin,
I get confused with the name ArticlePlugin, as it is not used as such.
