+++ src/article/converter/jats/createJatsExporter.js
@@ -1,10 +1,10 @@
 import TextureConfigurator from '../../../TextureConfigurator'
-import ArticlePlugin from '../../ArticlePlugin'
+import ArticlePackage from '../../ArticlePackage'
 
 // TODO: this is only needed for testing, so we should move this into test helpers
 export default function createJatsExporter (jatsDom, doc) {
   let config = new TextureConfigurator()
-  config.import(ArticlePlugin)
+  config.import(ArticlePackage)
   let articleConfig = config.getConfiguration('article')
   let exporter = articleConfig.createExporter('jats', doc)
   return exporter
