+++ src/article/converter/jats/createJatsImporter.js
@@ -1,12 +1,12 @@
 import { DocumentSchema } from 'substance'
 import TextureConfigurator from '../../../TextureConfigurator'
-import ArticlePlugin from '../../ArticlePlugin'
+import ArticlePackage from '../../ArticlePackage'
 import InternalArticleDocument from '../../InternalArticleDocument'
 
 // TODO: this is only needed for testing, so we should move this into test helpers
 export default function createJatsImporter (doc) {
   let config = new TextureConfigurator()
-  config.import(ArticlePlugin)
+  config.import(ArticlePackage)
   let articleConfig = config.getConfiguration('article')
   if (!doc) {
     let schema = new DocumentSchema({
