+++ src/article/index.js
@@ -8,7 +8,7 @@ export * from './ArticleConstants'
 export * from './articleHelpers'
 export { default as ArticleLoader } from './ArticleLoader'
 export { default as ArticlePanel } from './ArticlePanel'
-export { default as ArticlePlugin } from './ArticlePlugin'
+export { default as ArticlePackage } from './ArticlePackage'
 export { default as ArticleSerializer } from './ArticleSerializer'
 export { default as ArticleSession } from './ArticleSession'
 export { default as InternalArticleDocument } from './InternalArticleDocument'
